zonal stats enhancements
