+++ wradlib/tests/test_io.py
@@ -9,6 +9,7 @@ import subprocess
 import sys
 import tempfile
 import zlib
+from dataclasses import dataclass
 
 import deprecation
 import numpy as np
@@ -22,6 +23,7 @@ from . import (
     requires_dask,
     requires_data,
     requires_gdal,
+    requires_geos,
     requires_h5netcdf,
     requires_h5py,
     requires_netcdf,
@@ -1432,6 +1434,155 @@ class TestVector:
         geofile = util.get_wradlib_data_file(filename)
         io.gdal.open_vector(geofile)
         io.gdal.open_vector(geofile, "ESRI Shapefile")
+
+
+@pytest.fixture
+def data_source():
+    @dataclass(init=False, repr=False, eq=False)
+    class Data:
+        # create synthetic box
+        box0 = np.array(
+            [
+                [2600000.0, 5630000.0],
+                [2600000.0, 5640000.0],
+                [2610000.0, 5640000.0],
+                [2610000.0, 5630000.0],
+                [2600000.0, 5630000.0],
+            ]
+        )
+
+        box1 = np.array(
+            [
+                [2700000.0, 5630000.0],
+                [2700000.0, 5640000.0],
+                [2710000.0, 5640000.0],
+                [2710000.0, 5630000.0],
+                [2700000.0, 5630000.0],
+            ]
+        )
+
+        data = np.array([box0, box1], dtype=object)
+
+        ds = io.VectorSource(data)
+
+        values1 = np.array([47.11, 47.11])
+        values2 = np.array([47.11, 15.08])
+
+    yield Data
+
+
+@requires_geos
+class TestVectorSource:
+    @requires_data
+    @requires_gdal
+    def test__check_src(self):
+        from osgeo import osr
+
+        proj_gk2 = osr.SpatialReference()
+        proj_gk2.ImportFromEPSG(31466)
+        filename = util.get_wradlib_data_file("shapefiles/agger/agger_merge.shp")
+        assert len(io.VectorSource(filename, srs=proj_gk2).data) == 13
+
+    @requires_gdal
+    def test_error(self):
+        with pytest.raises(ValueError):
+            filename = util.get_wradlib_data_file("shapefiles/agger/agger_merge.shp")
+            io.VectorSource(filename)
+        with pytest.raises(RuntimeError):
+            io.VectorSource("test_zonalstats.py")
+
+    @requires_gdal
+    def test_data(self, data_source):
+        np.testing.assert_almost_equal(data_source.ds.data, data_source.data)
+
+    @requires_gdal
+    def test__get_data(self, data_source):
+        ds = io.VectorSource(data_source.data)
+        np.testing.assert_almost_equal(ds._get_data(), data_source.data)
+
+    @requires_gdal
+    def test_get_data_by_idx(self, data_source):
+        ds = io.VectorSource(data_source.data)
+        np.testing.assert_almost_equal(ds.get_data_by_idx([0]), data_source.data[0:1])
+        np.testing.assert_almost_equal(ds.get_data_by_idx([1]), data_source.data[1:2])
+        np.testing.assert_almost_equal(ds.get_data_by_idx([0, 1]), data_source.data)
+
+    @requires_gdal
+    def test_get_data_by_att(self, data_source):
+        ds = io.VectorSource(data_source.data)
+        np.testing.assert_almost_equal(
+            ds.get_data_by_att("index", 0), data_source.data[0:1]
+        )
+        np.testing.assert_almost_equal(
+            ds.get_data_by_att("index", 1), data_source.data[1:2]
+        )
+
+    @requires_gdal
+    def test_get_data_by_geom(self, data_source):
+        ds = io.VectorSource(data_source.data)
+        lyr = ds.ds.GetLayer()
+        lyr.ResetReading()
+        lyr.SetSpatialFilter(None)
+        lyr.SetAttributeFilter(None)
+        for i, feature in enumerate(lyr):
+            geom = feature.GetGeometryRef()
+            np.testing.assert_almost_equal(
+                ds.get_data_by_geom(geom), data_source.data[i : i + 1]
+            )
+
+    @requires_gdal
+    def test_set_attribute(self, data_source):
+        ds = io.VectorSource(data_source.data)
+        ds.set_attribute("test", data_source.values1)
+        assert np.allclose(ds.get_attributes(["test"]), data_source.values1)
+        ds.set_attribute("test", data_source.values2)
+        assert np.allclose(ds.get_attributes(["test"]), data_source.values2)
+
+    @requires_gdal
+    def test_get_attributes(self, data_source):
+        ds = io.VectorSource(data_source.data)
+        ds.set_attribute("test", data_source.values2)
+        assert ds.get_attributes(["test"], filt=("index", 0)) == data_source.values2[0]
+        assert ds.get_attributes(["test"], filt=("index", 1)) == data_source.values2[1]
+
+    @requires_data
+    @requires_gdal
+    def test_get_geom_properties(self):
+        from osgeo import osr
+
+        proj = osr.SpatialReference()
+        proj.ImportFromEPSG(31466)
+        filename = util.get_wradlib_data_file("shapefiles/agger/" "agger_merge.shp")
+        test = io.VectorSource(filename, proj)
+        np.testing.assert_array_equal(
+            [[76722499.98474795]], test.get_geom_properties(["Area"], filt=("FID", 1))
+        )
+
+    @requires_gdal
+    def test_dump_vector(self, data_source):
+        ds = io.VectorSource(data_source.data)
+        ds.dump_vector(tempfile.NamedTemporaryFile(mode="w+b").name)
+
+    @requires_data
+    @requires_gdal
+    def test_dump_raster(self):
+        from osgeo import osr
+
+        proj = osr.SpatialReference()
+        proj.ImportFromEPSG(31466)
+        filename = util.get_wradlib_data_file("shapefiles/agger/" "agger_merge.shp")
+        test = io.VectorSource(filename, srs=proj)
+        test.dump_raster(
+            tempfile.NamedTemporaryFile(mode="w+b").name,
+            driver="netCDF",
+            pixel_size=100.0,
+        )
+        test.dump_raster(
+            tempfile.NamedTemporaryFile(mode="w+b").name,
+            driver="netCDF",
+            pixel_size=100.0,
+            attr="FID",
+        )
 
 
 class TestIris:
